fix taskservice delete

+++ src/main/java/com/qa/todo/service/TaskService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import com.qa.todo.dto.TaskDTO;
 import com.qa.todo.exception.TaskNotFoundException;
+import com.qa.todo.exception.UserNotFoundException;
 import com.qa.todo.persistence.domain.Task;
 import com.qa.todo.persistence.repo.TaskRepository;
 import com.qa.todo.utils.SpringBeanUtils;
@@ -61,9 +62,14 @@ public class TaskService {
 	}
 	
 
-	public boolean delete(Long id) {
+	public Boolean delete(Long id) {
+		if(!this.repo.existsById(id)) {
+			throw new UserNotFoundException();
+		}
 		this.repo.deleteById(id);
-		return this.repo.existsById(id);
+		return !this.repo.existsById(id);
+//		this.repo.deleteById(id);
+//		return this.repo.existsById(id);
 	}
 	
 }
